custom 레포지토리 인터페이스화, 일반 레포지토리 상속받도록 수정
