+++ ecommerce/src/main/java/com/hello/ecommerce/repository/UsersRepository.java
@@ -1,0 +1,9 @@
+package com.hello.ecommerce.repository;
+
+import com.hello.ecommerce.entity.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UsersRepository extends JpaRepository<Users, Long> {
+}
